+++ iOS/MoEngageConfiguration.h
@@ -8,9 +8,10 @@
 //
 
 #define kMoEngageAppID @"Your App ID"
-#define kMoEngageRegion @"DEFAULT"  // DEFAULT / EU / SERV3
+#define kMoEngageDataCenter @"DATA_CENTER_01"  // DATA_CENTER_01 / DATA_CENTER_02 / DATA_CENTER_03
 #define kMoEngageLogsEnabled NO
 #define kDefaultIDFATrackingOptedOut NO
 #define kDefaultIDFVTrackingOptedOut NO
 #define kMoEngageUnityControllerSwizzlingEnabled YES
+#define kMoEngageEncryptNetworkRequests NO